+++ generator/src/org/intellij/grammar/generator/JavaParserGenerator.java
@@ -1839,7 +1839,7 @@ public final class JavaParserGenerator extends Generator {
     var generateTokenTypeConversions = G.generateTokenTypes && !mySimpleTokens.isEmpty();
     for (int i = 0; i < sortedCompositeTypesArr.length; i++) {
       String elementType = sortedCompositeTypesArr[i];
-      String elementTypeAccessor = "INSTANCE.get" + elementType.substring(0, 1).toUpperCase() + elementType.substring(1) + "()";
+      String elementTypeAccessor = "INSTANCE.get" + StringUtil.capitalize(elementType) + "()";
       out("new %s<%s, %s>(%s.%s, %s.%s)" + (i != sortedCompositeTypesArr.length - 1 || generateTokenTypeConversions ? "," : ""),
           shorten(KotlinBnfConstants.KT_PAIR_CLASS),
           shorten(KotlinBnfConstants.KT_ELEMENT_TYPE_CLASS), shorten(JavaBnfConstants.IELEMENTTYPE_CLASS),
@@ -1854,7 +1854,7 @@ public final class JavaParserGenerator extends Generator {
         String tokenName = ObjectUtils.chooseNotNull(mySimpleTokens.get(tokenText), tokenText);
         if (isIgnoredWhitespaceToken(tokenName, tokenText)) continue;
         var elementType = getElementType(tokenName);
-        String elementTypeAccessor = "INSTANCE.get" + elementType.substring(0, 1).toUpperCase() + elementType.substring(1) + "()";
+        String elementTypeAccessor = "INSTANCE.get" + StringUtil.capitalize(elementType) + "()";
         out("new %s<%s, %s>(%s.%s, %s.%s)" + (i != mySimpleTokensArr.length - 1 ? "," : ""),
             shorten(KotlinBnfConstants.KT_PAIR_CLASS),
             shorten(KotlinBnfConstants.KT_ELEMENT_TYPE_CLASS), shorten(JavaBnfConstants.IELEMENTTYPE_CLASS),